separate: meal class and util functions from main script
